+++ lib/components/features/event-registration/forms/SignedInWalletForm.tsx
@@ -1,8 +1,9 @@
 import { useForm, UseFormReturn } from "react-hook-form";
 import { useEffect } from "react";
+import { useAtom } from "jotai";
 import clsx from "clsx";
 
-import { buyerWalletAtom, formInstancesAtom, registrationModal, useAtom, useSetAtom } from "../store";
+import { buyerWalletAtom, formInstancesAtom, registrationModal, useSetAtom } from "../store";
 import { Button, LabeledInput, Menu, MenuItem, modal, toast } from "$lib/components/core";
 import { SetUserWalletDocument } from "$lib/graphql/generated/backend/graphql";
 import { formatWallet } from "$lib/utils/crypto";
@@ -18,11 +19,11 @@ export function SignedInWalletForm({ required }: { required: boolean }) {
 
   const setBuyerWallet = useSetAtom(buyerWalletAtom);
 
-  const userWallets = me?.wallets_new?.ethereum?.filter((wallet: string) => wallet !== me?.wallet_custodial) || [];
+  const userWallets = (me?.wallets_new?.ethereum ?? []).filter((wallet: string) => wallet && wallet !== me?.wallet_custodial);
 
   const form: UseFormReturn<{ selectedAddress: string }> = useForm<{ selectedAddress: string }>({
     defaultValues: {
-      selectedAddress: userWallets[0] || '',
+      selectedAddress: '',
     },
   });
   const setFormInstances = useSetAtom(formInstancesAtom);